arch/dts:init mods bplus ilim as 3A
For mods_b_plus en switch chipset, it has a ilim
pin to control current limitation. And we need to
init it as 3A limitation, which need to set the pin
of gpio_muc_bplus_disacharge as high. So we
add its initialization in dts to configure it as logic
"hihg".

Change-Id: I38a23a646272d6ec41e3d50c514de154fa18e0e0
Signed-off-by: weiweij <weiweij@motorola.com>
Reviewed-on: https://gerrit.mot.com/1106092
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key